Add an implementation of SchedulePage.vue
Currently missing connection to backend for sake of not pushing unstable code. Changes to router to accomodate changes, currently doesn't restrict access. Idea is that selectedCells contain all the information to change timeslot.avaliabilty++ of the correct timeslot for the algorithm. To implement is connection to backend and changing cells selected in previous sessions to red.